pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-ai"
version = "2.0.0"
description = "Advanced AI-powered network security analysis with multi-dataset ensemble learning and threat intelligence"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "NEXUS-AI Development Team", email = "team@nexus-ai.com"}
]
maintainers = [
    {name = "NEXUS-AI Development Team", email = "team@nexus-ai.com"}
]
keywords = ["security", "ai", "machine-learning", "network-analysis", "threat-intelligence", "cybersecurity", "advanced", "sophisticated"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Networking :: Monitoring",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "lightgbm>=3.3.0",
    "xgboost>=1.6.0",
    "transformers>=4.20.0",
    "shap>=0.40.0",
    "imbalanced-learn>=0.9.0",
    "pydantic>=1.9.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
    "xmltodict>=0.13.0",
    "colorama>=0.4.5",
    "rich>=12.0.0",
    "click>=8.1.0",
    "psutil>=5.8.0",
    "joblib>=1.1.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.10.0",
    "python-dotenv>=0.19.0",
    "httpx>=0.24.0",
    "aiohttp>=3.8.0",
    "ratelimit>=2.2.1",
    "diskcache>=5.2.0",
    "structlog>=22.1.0",
    "cryptography>=3.4.0",
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "flask-socketio>=5.3.0",
    "python-socketio>=5.8.0",
    "eventlet>=0.33.0",
    "dataclasses-json>=0.5.0",
    "ipaddress>=1.0.23",
    "asyncio-mqtt>=0.11.0",
    "prometheus-client>=0.17.0",
    "redis>=4.5.0",
    "celery>=5.3.0",
    "flower>=2.0.0",
    "sentry-sdk>=1.28.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation>=0.40b0",
    "opentelemetry-exporter-jaeger>=1.20.0",
    "opentelemetry-exporter-prometheus>=0.40b0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "markupsafe>=2.1.0",
    "itsdangerous>=2.1.0",
    "werkzeug>=2.3.0",
    "blinker>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-mock>=3.8.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "pre-commit>=2.19.0",
    "coverage>=7.2.0",
    "tox>=4.0.0",
    "pytest-xdist>=3.0.0",
    "pytest-html>=3.1.0",
    "pytest-benchmark>=4.0.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.17.0",
    "sphinx-autodoc-typehints>=1.19.0",
    "sphinx-copybutton>=0.5.0",
    "sphinx-tabs>=3.4.0",
]
docker = [
    "docker>=6.0.0",
    "docker-compose>=1.29.0",
]
monitoring = [
    "prometheus-client>=0.17.0",
    "grafana-api>=1.0.3",
    "elasticsearch>=8.0.0",
    "kibana>=8.0.0",
    "logstash>=8.0.0",
]
ml = [
    "tensorflow>=2.12.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.20.0",
    "datasets>=2.12.0",
    "accelerate>=0.20.0",
    "optuna>=3.2.0",
    "mlflow>=2.5.0",
    "wandb>=0.15.0",
]
security = [
    "cryptography>=3.4.0",
    "bcrypt>=4.0.0",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.6",
    "argon2-cffi>=21.3.0",
    "certifi>=2023.7.22",
    "urllib3>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/Dreepsy/Project_Nexus"
Documentation = "https://github.com/Dreepsy/Project_Nexus#readme"
Repository = "https://github.com/Dreepsy/Project_Nexus.git"
"Bug Tracker" = "https://github.com/Dreepsy/Project_Nexus/issues"
"Changelog" = "https://github.com/Dreepsy/Project_Nexus/blob/main/CHANGELOG.md"

[project.scripts]
nexus-ai = "nexus.cli.commands:main"
NexusAI = "nexus.cli.commands:main"

[project.gui-scripts]
nexus-ai-gui = "nexus.cli.gui:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
nexus = ["*.yaml", "*.yml", "*.json", "*.md", "templates/*.html", "static/*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["nexus"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "performance: marks tests as performance tests",
    "security: marks tests as security tests",
]

[tool.coverage.run]
source = ["src/nexus"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "docs"]
skips = ["B101", "B601"]

[tool.safety]
output = "json"

[tool.poetry]
name = "nexus-ai"
version = "2.0.0"
description = "Advanced AI-powered network security analysis with multi-dataset ensemble learning and threat intelligence"
authors = ["NEXUS-AI Development Team <team@nexus-ai.com>"]
readme = "README.md"
packages = [{include = "nexus", from = "src"}]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Networking :: Monitoring",
    "Topic :: Software Development :: Libraries :: Python Modules",
] 